{test.softwareC => driver/software.c}anvas
This is the main step to make the `driver/software` package independent
from the `test` package.
This step introduces a circular dependency which will be resolved with
the next commit.

piggy-backed: Move #doResize() to the correct place.
