+++ driver/software/canvas_test.go
@@ -1,4 +1,4 @@
-package test_test
+package software_test
 
 import (
 	"image/color"
@@ -7,13 +7,13 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/test"
+	"fyne.io/fyne/v2/driver/software"
 	"fyne.io/fyne/v2/theme"
 	"fyne.io/fyne/v2/widget"
 )
 
 func Test_canvas_Capture(t *testing.T) {
-	c := test.NewCanvas()
+	c := software.NewCanvas()
 	c.Resize(fyne.NewSquareSize(100))
 
 	img := c.Capture()
@@ -29,7 +29,7 @@ func Test_canvas_Capture(t *testing.T) {
 }
 
 func Test_canvas_InteractiveArea(t *testing.T) {
-	c := test.NewCanvas()
+	c := software.NewCanvas()
 	// TODO a custom interactive area that does not break other tests (overlays restrict to the space)
 	c.Resize(fyne.NewSize(600, 400))
 	pos, size := c.InteractiveArea()
@@ -38,7 +38,7 @@ func Test_canvas_InteractiveArea(t *testing.T) {
 }
 
 func Test_canvas_PixelCoordinateAtPosition(t *testing.T) {
-	c := test.NewCanvas()
+	c := software.NewCanvas()
 
 	pos := fyne.NewPos(4, 4)
 	c.SetScale(2.5)
@@ -48,7 +48,7 @@ func Test_canvas_PixelCoordinateAtPosition(t *testing.T) {
 }
 
 func Test_canvas_TransparentCapture(t *testing.T) {
-	c := test.NewTransparentCanvasWithPainter(nil)
+	c := software.NewTransparentCanvasWithPainter(nil)
 	c.Resize(fyne.NewSquareSize(10))
 
 	img := c.Capture()
@@ -64,7 +64,7 @@ func Test_canvas_TransparentCapture(t *testing.T) {
 }
 
 func Test_canvas_Resize(t *testing.T) {
-	c := test.NewCanvas()
+	c := software.NewCanvas()
 	assert.Equal(t, float32(100), c.Size().Width) // backwards compatible
 
 	smallSize := fyne.NewSize(10, 5)
@@ -74,7 +74,7 @@ func Test_canvas_Resize(t *testing.T) {
 	assert.Greater(t, c.Size().Width, smallSize.Width)
 	assert.Greater(t, c.Size().Height, smallSize.Height)
 
-	c = test.NewCanvas()
+	c = software.NewCanvas()
 	largeSize := fyne.NewSize(100, 75)
 	c.Resize(largeSize)
 	c.SetContent(widget.NewLabel("Smaller"))
